Reorganize patchers
