+++ lib/mix/tasks/surface/surface.init/command.ex
@@ -1,0 +1,4 @@
+defmodule Mix.Tasks.Surface.Init.Command do
+  @callback file_patchers(assigns :: map()) :: [map()]
+  @callback create_files(assigns :: map()) :: list()
+end
