+++ test/mix/tasks/surface/surface.init/commands/common_test.exs
@@ -1,0 +1,123 @@
+defmodule Mix.Tasks.Surface.Init.Commands.CommonTest do
+  use ExUnit.Case, async: true
+
+  alias Mix.Tasks.Surface.Init.Patcher
+  import Mix.Tasks.Surface.Init.Commands.Common
+
+  describe "add_surface_live_reload_pattern_to_endpoint_config" do
+    test "update live_reload patterns" do
+      code = """
+      import Config
+
+      # Watch static and templates for browser reloading.
+      config :my_app, MyAppWeb.Endpoint,
+        reloadable_compilers: [:phoenix, :elixir, :surface],
+        live_reload: [
+          patterns: [
+            ~r"lib/my_app_web/(live|views)/.*(ex)$",
+            ~r"lib/my_app_web/templates/.*(eex)$"
+          ]
+        ]
+      """
+
+      {:patched, updated_code} =
+        Patcher.patch_code(
+          code,
+          add_surface_live_reload_pattern_to_endpoint_config(:my_app, MyAppWeb, "lib/my_app_web")
+        )
+
+      assert updated_code == """
+             import Config
+
+             # Watch static and templates for browser reloading.
+             config :my_app, MyAppWeb.Endpoint,
+               reloadable_compilers: [:phoenix, :elixir, :surface],
+               live_reload: [
+                 patterns: [
+                   ~r"lib/my_app_web/(live|views|components)/.*(ex|sface|js)$",
+                   ~r"lib/my_app_web/templates/.*(eex)$"
+                 ]
+               ]
+             """
+    end
+
+    test "don't apply it if already patched" do
+      code = """
+      import Config
+
+      # Watch static and templates for browser reloading.
+      config :my_app, MyAppWeb.Endpoint,
+        reloadable_compilers: [:phoenix, :elixir, :surface],
+        live_reload: [
+          patterns: [
+            ~r"lib/my_app_web/(live|views|components)/.*(ex|sface|js)$",
+            ~r"lib/my_app_web/templates/.*(eex)$"
+          ]
+        ]
+      """
+
+      assert {:already_patched, ^code} =
+               Patcher.patch_code(
+                 code,
+                 add_surface_live_reload_pattern_to_endpoint_config(:my_app, MyAppWeb, "lib/my_app_web")
+               )
+    end
+  end
+
+  describe "add_import_surface_to_view_macro" do
+    test "add import Surface" do
+      code = """
+      defmodule MyAppWeb do
+        def view do
+          quote do
+            use Phoenix.View,
+              root: "lib/my_app_web/templates",
+              namespace: MyAppWeb
+
+            # Include shared imports and aliases for views
+            unquote(view_helpers())
+          end
+        end
+      end
+      """
+
+      {:patched, updated_code} = Patcher.patch_code(code, add_import_surface_to_view_macro(MyAppWeb))
+
+      assert updated_code == """
+             defmodule MyAppWeb do
+               def view do
+                 quote do
+                   use Phoenix.View,
+                     root: "lib/my_app_web/templates",
+                     namespace: MyAppWeb
+
+                   # Include shared imports and aliases for views
+                   unquote(view_helpers())
+                   import Surface
+                 end
+               end
+             end
+             """
+    end
+
+    test "don't apply it if already patched" do
+      code = """
+      defmodule MyAppWeb do
+        def view do
+          quote do
+            use Phoenix.View,
+              root: "lib/my_app_web/templates",
+              namespace: MyAppWeb
+
+            # Include shared imports and aliases for views
+            unquote(view_helpers())
+            import Surface
+          end
+        end
+      end
+      """
+
+      assert {:already_patched, ^code} = Patcher.patch_code(code, add_import_surface_to_view_macro(MyAppWeb))
+    end
+  end
+end
